refactor: merge describe into info — linch info entity/action/relation
Consolidate linch describe subcommands into linch info:
- linch info             — project overview (existing)
- linch info entity <n>  — detailed entity info (from describe)
- linch info action <n>  — detailed action info (from describe)
- linch info relation <n> — detailed relation info (from describe)
All support --json for AI tool consumption.
